requirements.txt
streamlit
pandas
scikit-learn
joblib
numpy
groq
python-dotenv